(db update required, see below or v5.0 release notes) Changed Tag.id type from Long to Integer. Changed NewsItem.id type from Long to Integer. alter table tags change column id id integer not null auto_increment; alter table projects_related_to_tags change column tag_fk tag_fk integer not null; alter table newsitem change column id id integer not null auto_increment;

+++ src/main/java/org/wise/portal/dao/project/TagDao.java
@@ -41,10 +41,10 @@ public interface TagDao<T extends Tag> extends SimpleDao<T> {
 	Tag getTagByName(String name);
 	
 	/**
-	 * Given a <code>Long</code> tagId, removes that tag from the database
+	 * Given a <code>Integer</code> tagId, removes that tag from the database
 	 * if it is not used by any projects.
 	 * 
-	 * @param Long - tag id
+	 * @param Integer - tag id
 	 */
-	void removeIfOrphaned(Long tagId);
+	void removeIfOrphaned(Integer tagId);
 }
